refactor: main문을 LottoController로 이동
